requirements.txt
streamlit==1.26.0
numpy==1.21.3
Pillow==8.4.0
pandas==2.1.1
altair==5.1.1
folium==0.14.0
streamlit-folium==0.14.0
